Cleanup sergios comments
