+++ inductiva/projects/__init__.py
@@ -1,2 +1,3 @@
 #pylint: disable=missing-module-docstring
-from inductiva.projects.project import Project, get_current_project
+from inductiva.projects.project import (Project, get_current_project,
+                                        get_projects)
